Bug in fB2 include test squashed
Original version included the requirement fhSamples(fh) > 2. This made it
impossible for the fB2 component to be included in models with only 2
samples. This is a wrong of course, it should be included whenever there
are at least 2 (or > 1).

This only rarely impacted our data, since we usually set the voltage high
enough that the B2 peak is off-scale anyways. However, for samples where
the B2 peak *should* be included, this bug would make the model fitting
fail, or give inaccurate results. The loss of accuracy is usually
relatively small (1-2% in cases I looked at). It looks like when the loss
in accuracy would be large enough to matter, the fitting just fails
outright.